server/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==1.16.0
dominate==2.6.0
eventlet==0.30.2
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
psycopg2
pyparsing==3.0.7
python-dotenv==0.20.0
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
